backend/requirements.txt
fastapi
uvicorn
python-docx
python-pptx
PyPDF2
openai
pydantic
python-multipart
pyinstaller
